requirements.txt
Flask
pandas
scikit-learn
PyPDF2
python-docx
flask-cors
yagmail